Deprecate existing kernel connection
