+++ SUMMARY.md
@@ -6,7 +6,6 @@
   * [Getting Started](docs/Usage/GettingStarted.md)
   * [Examples](docs/Usage/Examples.md)
   * [Remote Kernels](docs/Usage/RemoteKernelConnection.md)
-  * [Custom Kernel Connection (deprecated)](docs/Usage/CustomKernelConnection.md)
 * [Troubleshooting Guide](docs/Troubleshooting.md)
 * [Style Customization](docs/StyleCustomization.md)
 * [Plugin API](docs/PluginAPI.md)
