+++ lib/kernel-manager.js
@@ -11,7 +11,7 @@ import ZMQKernel from "./zmq-kernel";
 
 import KernelPicker from "./kernel-picker";
 import store from "./store";
-import { getEditorDirectory, log, deprecationNote } from "./utils";
+import { getEditorDirectory, log } from "./utils";
 
 import type { Connection } from "./zmq-kernel";
 
@@ -27,26 +27,6 @@ class KernelManager {
     editor: atom$TextEditor,
     onStarted: (kernel: ZMQKernel) => void
   ) {
-    try {
-      const rootDirectory = editor
-        ? getEditorDirectory(editor)
-        : atom.project.getPaths()[0];
-
-      const connectionFile = path.join(
-        rootDirectory,
-        "hydrogen",
-        "connection.json"
-      );
-      const connectionString = fs.readFileSync(connectionFile, "utf8");
-      const connection = JSON.parse(connectionString);
-      this.startExistingKernel(grammar, connection, connectionFile, onStarted);
-      return;
-    } catch (e) {
-      if (e.code !== "ENOENT") {
-        console.error("KernelManager: Cannot start existing kernel:\n", e);
-      }
-    }
-
     this.getKernelSpecForGrammar(grammar, kernelSpec => {
       if (!kernelSpec) {
         const message = `No kernel for grammar \`${grammar.name}\` found`;
@@ -57,46 +37,6 @@ class KernelManager {
       }
 
       this.startKernel(kernelSpec, grammar, onStarted);
-    });
-  }
-
-  startExistingKernel(
-    grammar: atom$Grammar,
-    connection: Connection,
-    connectionFile: string,
-    onStarted: (kernel: ZMQKernel) => void
-  ) {
-    deprecationNote();
-    const language = grammar.name;
-
-    log("KernelManager: startExistingKernel: Assuming", language);
-
-    const kernelSpec = {
-      display_name: "Existing Kernel",
-      language,
-      argv: [],
-      env: {}
-    };
-
-    const kernel = new ZMQKernel(
-      kernelSpec,
-      grammar,
-      connection,
-      connectionFile
-    );
-
-    kernel.connect(() => {
-      log(
-        "KernelManager: startExistingKernelFor: existing kernel for",
-        language,
-        "connected"
-      );
-
-      store.newKernel(kernel);
-
-      this._executeStartupCode(kernel);
-
-      if (onStarted) onStarted(kernel);
     });
   }
 
@@ -132,39 +72,10 @@ class KernelManager {
       stdio: ["ignore", "pipe", "pipe"]
     };
 
-    launchSpec(
-      kernelSpec,
-      options
-    ).then(({ config, connectionFile, spawn }) => {
-      const kernel = new ZMQKernel(
-        kernelSpec,
-        grammar,
-        config,
-        connectionFile,
-        spawn,
-        options
-      );
-
-      kernel.connect(() => {
-        log("KernelManager: startKernel:", displayName, "connected");
-
-        store.newKernel(kernel);
-
-        this._executeStartupCode(kernel);
-
-        if (onStarted) onStarted(kernel);
-      });
+    const kernel = new ZMQKernel(kernelSpec, grammar, options, () => {
+      store.newKernel(kernel);
+      if (onStarted) onStarted(kernel);
     });
-  }
-
-  _executeStartupCode(kernel: ZMQKernel) {
-    const displayName = kernel.kernelSpec.display_name;
-    let startupCode = Config.getJson("startupCode")[displayName];
-    if (startupCode) {
-      log("KernelManager: Executing startup code:", startupCode);
-      startupCode = `${startupCode} \n`;
-      kernel.execute(startupCode);
-    }
   }
 
   getAllKernelSpecs(callback: Function) {
