+++ lib/utils.js
@@ -6,8 +6,6 @@ import ReactDOM from "react-dom";
 import _ from "lodash";
 import os from "os";
 import path from "path";
-// $FlowFixMe
-import { shell } from "electron";
 
 import Config from "./config";
 import store from "./store";
@@ -134,32 +132,6 @@ export function renderDevTools() {
       log("Could not enable dev tools", e);
     }
   }
-}
-
-export function deprecationNote() {
-  atom.notifications.addWarning("This feature will be deprecated soon!", {
-    description:
-      "Connecting to existing kernels via a `connection.json` file will be deprecated soon.\n\nFor some time now Hydrogen supports using [kernel gateways](https://nteract.gitbooks.io/hydrogen/docs/Usage/RemoteKernelConnection.html) for connection to existing kernels. Using that option is a lot simpler yet very powerful.\n\nPlease get in touch with us if using remote kernels isn't a option for you.",
-    dismissable: true,
-    buttons: [
-      {
-        className: "icon icon-x",
-        text: "I really need this feature",
-        onDidClick: () => {
-          shell.openExternal("https://github.com/nteract/hydrogen/issues/858");
-        }
-      },
-      {
-        className: "icon icon-check",
-        text: "I'll try remote kernels",
-        onDidClick: () => {
-          shell.openExternal(
-            "https://nteract.gitbooks.io/hydrogen/docs/Usage/RemoteKernelConnection.html"
-          );
-        }
-      }
-    ]
-  });
 }
 
 export function hotReloadPackage() {
