+++ lib/zmq-kernel.js
@@ -3,8 +3,9 @@
 import fs from "fs";
 import { Message, Socket } from "jmp";
 import v4 from "uuid/v4";
-import { launchSpecFromConnectionInfo } from "spawnteract";
+import { launchSpec, launchSpecFromConnectionInfo } from "spawnteract";
 
+import Config from "./config";
 import Kernel from "./kernel";
 import InputView from "./input-view";
 import { log } from "./utils";
@@ -37,43 +38,28 @@ export default class ZMQKernel extends Kernel {
   constructor(
     kernelSpec: Kernelspec,
     grammar: atom$Grammar,
-    connection: Connection,
-    connectionFile: string,
-    kernelProcess: ?child_process$ChildProcess,
-    options: ?Object
+    options: Object,
+    onStarted: ?Function
   ) {
     super(kernelSpec, grammar);
-    this.connection = connection;
-    this.connectionFile = connectionFile;
     this.options = options || {};
 
-    if (kernelProcess) {
-      this.kernelProcess = kernelProcess;
-      log("ZMQKernel: @kernelProcess:", this.kernelProcess);
+    launchSpec(
+      kernelSpec,
+      options
+    ).then(({ config, connectionFile, spawn }) => {
+      this.connection = config;
+      this.connectionFile = connectionFile;
+      this.kernelProcess = spawn;
 
-      this.kernelProcess.stdout.on("data", (data: string | Buffer) => {
-        data = data.toString();
+      this.monitorNotifications(spawn);
 
-        if (atom.config.get("Hydrogen.kernelNotifications")) {
-          atom.notifications.addInfo(this.kernelSpec.display_name, {
-            description: data,
-            dismissable: true
-          });
-        } else {
-          log("ZMQKernel: stdout:", data);
-        }
+      this.connect(() => {
+        this._executeStartupCode();
+
+        if (onStarted) onStarted(this);
       });
-
-      this.kernelProcess.stderr.on("data", (data: string | Buffer) => {
-        atom.notifications.addError(this.kernelSpec.display_name, {
-          description: data.toString(),
-          dismissable: true
-        });
-      });
-    } else {
-      log("ZMQKernel: connectionFile:", this.connectionFile);
-      atom.notifications.addInfo("Using an existing kernel connection");
-    }
+    });
   }
 
   connect(done: ?Function) {
@@ -103,6 +89,28 @@ export default class ZMQKernel extends Kernel {
     this.stdinSocket.on("message", this.onStdinMessage.bind(this));
 
     this.monitor(done);
+  }
+
+  monitorNotifications(childProcess: child_process$ChildProcess) {
+    childProcess.stdout.on("data", (data: string | Buffer) => {
+      data = data.toString();
+
+      if (atom.config.get("Hydrogen.kernelNotifications")) {
+        atom.notifications.addInfo(this.kernelSpec.display_name, {
+          description: data,
+          dismissable: true
+        });
+      } else {
+        log("ZMQKernel: stdout:", data);
+      }
+    });
+
+    childProcess.stderr.on("data", (data: string | Buffer) => {
+      atom.notifications.addError(this.kernelSpec.display_name, {
+        description: data.toString(),
+        dismissable: true
+      });
+    });
   }
 
   monitor(done: ?Function) {
@@ -142,22 +150,24 @@ export default class ZMQKernel extends Kernel {
       atom.notifications.addWarning("Cannot interrupt this kernel", {
         detail: "Kernel interruption is currently not supported in Windows."
       });
-    } else if (this.kernelProcess) {
+    } else {
       log("ZMQKernel: sending SIGINT");
       this.kernelProcess.kill("SIGINT");
-    } else {
-      log("ZMQKernel: cannot interrupt an existing kernel");
-      atom.notifications.addWarning("Cannot interrupt an existing kernel");
     }
   }
 
   _kill() {
-    if (this.kernelProcess) {
-      log("ZMQKernel: sending SIGKILL");
-      this.kernelProcess.kill("SIGKILL");
-    } else {
-      log("ZMQKernel: cannot kill an existing kernel");
-      atom.notifications.addWarning("Cannot kill this kernel");
+    log("ZMQKernel: sending SIGKILL");
+    this.kernelProcess.kill("SIGKILL");
+  }
+
+  _executeStartupCode() {
+    const displayName = this.kernelSpec.display_name;
+    let startupCode = Config.getJson("startupCode")[displayName];
+    if (startupCode) {
+      log("KernelManager: Executing startup code:", startupCode);
+      startupCode = `${startupCode} \n`;
+      this.execute(startupCode);
     }
   }
 
@@ -171,28 +181,22 @@ export default class ZMQKernel extends Kernel {
   }
 
   restart(onRestarted: ?Function) {
-    if (this.kernelProcess) {
-      if (this.executionState === "restarting") {
-        return;
-      }
-      this.setExecutionState("restarting");
-      this.shutdown(true);
-      this._kill();
-      const { spawn } = launchSpecFromConnectionInfo(
-        this.kernelSpec,
-        this.connection,
-        this.connectionFile,
-        this.options
-      );
-      this.kernelProcess = spawn;
-      this.monitor(() => {
-        if (onRestarted) onRestarted(this);
-      });
-    } else {
-      log("ZMQKernel: restart ignored:", this);
-      atom.notifications.addWarning("Cannot restart this kernel");
-      if (onRestarted) onRestarted(this);
+    if (this.executionState === "restarting") {
+      return;
     }
+    this.setExecutionState("restarting");
+    this.shutdown(true);
+    this._kill();
+    const { spawn } = launchSpecFromConnectionInfo(
+      this.kernelSpec,
+      this.connection,
+      this.connectionFile,
+      this.options
+    );
+    this.kernelProcess = spawn;
+    this.monitor(() => {
+      if (onRestarted) onRestarted(this);
+    });
   }
 
   // onResults is a callback that may be called multiple times
@@ -435,10 +439,8 @@ export default class ZMQKernel extends Kernel {
 
     this.shutdown();
 
-    if (this.kernelProcess) {
-      this._kill();
-      fs.unlinkSync(this.connectionFile);
-    }
+    this._kill();
+    fs.unlinkSync(this.connectionFile);
 
     this.shellSocket.close();
     this.controlSocket.close();
